backend/requirements.txt
# Web框架
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
gunicorn==20.1.0

# 数据库
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
py2neo==2021.2.3
pymilvus==2.2.8

# 大语言模型和向量计算
openai==1.3.0
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
transformers==4.30.2

# 数据处理
pandas==1.5.3
scikit-learn==1.2.2

# 工具库
python-dotenv==1.0.0
requests==2.28.2
PyJWT==2.6.0
python-jose==3.3.0
passlib==1.7.4

# 日志和监控
logging==0.4.9.6
